Rename lr_r2 to lr2
