+++ unittests/files/opt/seagate/sspl/setup/resource_map/lr2/encl_api_response.py
